requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astrapy==0.7.7
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.15
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.2
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.9
effdet==0.4.1
emoji==2.14.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
fastapi==0.115.2
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
FlagEmbedding==1.2.11
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
geomet==0.2.1.post1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.136
layoutparser==0.3.4
llama-cloud==0.1.4
llama-index==0.11.19
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.19
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.15
llama-index-multi-modal-llms-openai==0.2.2
llama-index-postprocessor-flag-embedding-reranker==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-index-vector-stores-astra==0.1.2
llama-index-vector-stores-postgres==0.2.6
llama-parse==0.5.10
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
openai==1.52.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.9
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.13.2
pexpect==4.9.0
pgvector==0.2.5
pi_heif==0.20.0
pikepdf==9.3.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
posthog==3.7.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
pypandoc==1.14
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.0
unstructured-client==0.26.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
